+++ src/main/java/searchengine/dto/statistics/TotalStatistics.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class TotalStatistics {
-    private int sites;
-    private int pages;
-    private int lemmas;
+    private long sites;
+    private long pages;
+    private long lemmas;
     private boolean indexing;
 }
